добавил нотификацию при синхронизации данных
