+++ app/src/main/java/money/android/bignerdranch/com/moneytracker/UI/adapters/CategoriesSpinnerAdapter.java
@@ -22,9 +22,7 @@ import money.android.bignerdranch.com.moneytracker.entitys.CategoryEntity;
 
 public class CategoriesSpinnerAdapter extends ArrayAdapter<CategoryEntity> implements SpinnerAdapter{
 
-
     List<CategoryEntity> categories;
-
     public CategoriesSpinnerAdapter(Context context, List<CategoryEntity> categories) {
         super(context, 0, categories);
         this.categories = categories;
