+++ app/src/main/java/money/android/bignerdranch/com/moneytracker/UI/utils/ScrollingFABBehavior.java
@@ -2,38 +2,36 @@ package money.android.bignerdranch.com.moneytracker.UI.utils;
 
 import android.content.Context;
 import android.support.design.widget.CoordinatorLayout;
- import android.support.design.widget.FloatingActionButton;
+import android.support.design.widget.FloatingActionButton;
 import android.support.v4.view.ViewCompat;
-import android.support.v7.widget.RecyclerView;
- import android.util.AttributeSet;
- import android.view.View;
+import android.util.AttributeSet;
+import android.view.View;
 import android.view.animation.LinearInterpolator;
 
 
 public class ScrollingFABBehavior extends FloatingActionButton.Behavior {
 
-     private static final String TAG = "ScrollingFABBehavior";
-     public ScrollingFABBehavior(Context context, AttributeSet attrs) {
-         super();
-     }
+    public ScrollingFABBehavior(Context context, AttributeSet attrs) {
+        super();
+    }
 
-     @Override
-     public void onNestedScroll(CoordinatorLayout coordinatorLayout, FloatingActionButton child, View target, int dxConsumed, int dyConsumed, int dxUnconsumed, int dyUnconsumed) {
-         super.onNestedScroll(coordinatorLayout, child, target, dxConsumed, dyConsumed, dxUnconsumed, dyUnconsumed);
+    @Override
+    public void onNestedScroll(CoordinatorLayout coordinatorLayout, FloatingActionButton child, View target, int dxConsumed, int dyConsumed, int dxUnconsumed, int dyUnconsumed) {
+        super.onNestedScroll(coordinatorLayout, child, target, dxConsumed, dyConsumed, dxUnconsumed, dyUnconsumed);
 
-         //child -> Floating Action Button
-         if (dyConsumed > 0) {
-             CoordinatorLayout.LayoutParams layoutParams = (CoordinatorLayout.LayoutParams) child.getLayoutParams();
-             int fab_bottomMargin = layoutParams.bottomMargin;
-             child.animate().translationY(child.getHeight() + fab_bottomMargin).setInterpolator(new LinearInterpolator()).start();
-         } else if (dyConsumed < 0) {
-             child.animate().translationY(0).setInterpolator(new LinearInterpolator()).start();
-         }
-     }
+        //child -> Floating Action Button
+        if (dyConsumed > 0) {
+            CoordinatorLayout.LayoutParams layoutParams = (CoordinatorLayout.LayoutParams) child.getLayoutParams();
+            int fab_bottomMargin = layoutParams.bottomMargin;
+            child.animate().translationY(child.getHeight() + fab_bottomMargin).setInterpolator(new LinearInterpolator()).start();
+        } else if (dyConsumed < 0) {
+            child.animate().translationY(0).setInterpolator(new LinearInterpolator()).start();
+        }
+    }
 
-     @Override
-     public boolean onStartNestedScroll(CoordinatorLayout coordinatorLayout, FloatingActionButton child, View directTargetChild, View target, int nestedScrollAxes) {
-         return nestedScrollAxes == ViewCompat.SCROLL_AXIS_VERTICAL;
-     }
+    @Override
+    public boolean onStartNestedScroll(CoordinatorLayout coordinatorLayout, FloatingActionButton child, View directTargetChild, View target, int nestedScrollAxes) {
+        return nestedScrollAxes == ViewCompat.SCROLL_AXIS_VERTICAL;
+    }
 
- }
+}
