+++ app/src/main/java/money/android/bignerdranch/com/moneytracker/entitys/ExpensesEntity.java
@@ -5,8 +5,6 @@ import com.activeandroid.annotation.Column;
 import com.activeandroid.annotation.Table;
 import com.activeandroid.query.Delete;
 import com.activeandroid.query.Select;
-import com.activeandroid.query.Sqlable;
-
 import java.util.List;
