+++ app/src/main/java/money/android/bignerdranch/com/moneytracker/sync/TrackerSyncAdapter.java
@@ -2,44 +2,55 @@ package money.android.bignerdranch.com.moneytracker.sync;
 
 import android.accounts.Account;
 import android.accounts.AccountManager;
+import android.app.NotificationManager;
 import android.content.AbstractThreadedSyncAdapter;
-import android.content.ComponentName;
 import android.content.ContentProviderClient;
 import android.content.ContentResolver;
 import android.content.Context;
-import android.content.ServiceConnection;
+import android.content.SharedPreferences;
 import android.content.SyncRequest;
 import android.content.SyncResult;
+import android.graphics.BitmapFactory;
+import android.graphics.Color;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.IBinder;
-import android.util.Log;
-
+import android.provider.Settings;
+import android.support.v4.app.NotificationCompat;
+import android.support.v7.preference.PreferenceManager;
 import com.google.gson.Gson;
-
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
 import money.android.bignerdranch.com.moneytracker.R;
-import money.android.bignerdranch.com.moneytracker.UI.MainActivity;
 import money.android.bignerdranch.com.moneytracker.UI.utils.MoneyTrackerAplication;
 import money.android.bignerdranch.com.moneytracker.entitys.CategoryEntity;
 import money.android.bignerdranch.com.moneytracker.entitys.ExpensesEntity;
 import money.android.bignerdranch.com.moneytracker.rest.Models.CategoryModel;
 import money.android.bignerdranch.com.moneytracker.rest.Models.ExpensesModel;
-import money.android.bignerdranch.com.moneytracker.rest.Models.UserGetDataModel;
 import money.android.bignerdranch.com.moneytracker.rest.Models.UserSyncCategoriesModel;
 import money.android.bignerdranch.com.moneytracker.rest.Models.UserSyncExpensesModel;
 import money.android.bignerdranch.com.moneytracker.rest.RestService;
-import money.android.bignerdranch.com.moneytracker.services.ServiceSample;
-
 
 public class TrackerSyncAdapter extends AbstractThreadedSyncAdapter {
 
-
-
+    private static final long[] VIBRATE_PATTERN = new long[]{0, 100, 0, 200};
+    private static final int LED_LIGHTS_TIME_ON = 200;
+    private static final int LED_LIGHTS_TIME_OFF = 1500;
+    private static final int NOTIFICATION_ID = 4004;
+    private NotificationManager mNotificationManager;
+    private boolean isNotificationsEnabled;
+    private boolean isVibrateEnabled;
+    private boolean isSoundEnabled;
+    private boolean isLedEnabled;
+    private boolean isForegroundEnabled;
+    private SharedPreferences mSharedPreferences;
+    private String globalNotificationsKey;
+    private String vibrateNotificationsKey;
+    private String soundNotificationsKey;
+    private String ledNotificationsKey;
+    private String foregroundNotificationsKey;
+    private static final boolean DEFAULT_VALUE = true;
 
     public TrackerSyncAdapter(Context context, boolean autoInitialize) {
         super(context, autoInitialize);
@@ -50,13 +61,27 @@ public class TrackerSyncAdapter extends AbstractThreadedSyncAdapter {
         try {
             syncCategories();
             syncExpenses();
-
-
-        } catch (Exception e){
+            sendNotification();
+        } catch (Exception e) {
 
         }
     }
 
+    private void init() {
+        SharedPreferences mSharedPreferences = PreferenceManager.getDefaultSharedPreferences(getContext());
+
+        globalNotificationsKey = getContext().getString(R.string.pref_enable_notifications_key);
+        vibrateNotificationsKey = getContext().getString(R.string.pref_enable_vibrate_notifications_key);
+        soundNotificationsKey = getContext().getString(R.string.pref_enable_sound_notifications_key);
+        ledNotificationsKey = getContext().getString(R.string.pref_enable_led_notifications_key);
+        foregroundNotificationsKey = getContext().getString(R.string.pref_enable_foreground_notifications_key);
+
+        isNotificationsEnabled = mSharedPreferences.getBoolean(globalNotificationsKey, DEFAULT_VALUE);
+        isVibrateEnabled = mSharedPreferences.getBoolean(vibrateNotificationsKey, DEFAULT_VALUE);
+        isSoundEnabled = mSharedPreferences.getBoolean(soundNotificationsKey, DEFAULT_VALUE);
+        isLedEnabled = mSharedPreferences.getBoolean(ledNotificationsKey, DEFAULT_VALUE);
+        isForegroundEnabled = mSharedPreferences.getBoolean(foregroundNotificationsKey, DEFAULT_VALUE);
+    }
 
     private static void syncImmediately(Context context) {
         Bundle bundle = new Bundle();
@@ -79,7 +104,7 @@ public class TrackerSyncAdapter extends AbstractThreadedSyncAdapter {
     }
 
     private static void onAccountCreated(Account newAccount, Context context) {
-        final int SYNC_INTERVAL = 60;
+        final int SYNC_INTERVAL = 60 * 60 * 24;
         final int SYNC_FLEXTIME = SYNC_INTERVAL / 3;
         TrackerSyncAdapter.configurePeriodicSync(context, SYNC_INTERVAL, SYNC_FLEXTIME);
         ContentResolver.setSyncAutomatically(newAccount, context.getString(R.string.content_authority), true);
@@ -87,12 +112,12 @@ public class TrackerSyncAdapter extends AbstractThreadedSyncAdapter {
         syncImmediately(context);
     }
 
-    private void syncCategories (){
+    private void syncCategories() {
 
         RestService restService = new RestService();
         List<CategoryEntity> categoryEntityList = CategoryEntity.selectAll("");
         List<CategoryModel> categoryModels = new ArrayList<>();
-        for (int i = 0; i < categoryEntityList.size(); i ++) {
+        for (int i = 0; i < categoryEntityList.size(); i++) {
             CategoryModel categoryModel = new CategoryModel();
             categoryModel.setId(categoryEntityList.get(i).getId().toString());
             categoryModel.setTitle(categoryEntityList.get(i).getName());
@@ -103,18 +128,17 @@ public class TrackerSyncAdapter extends AbstractThreadedSyncAdapter {
 
         try {
             UserSyncCategoriesModel userSyncCategoriesModel = restService.userSyncCategoriesModel(gson, MoneyTrackerAplication.getAuthToken(), MoneyTrackerAplication.getGoogleAuthToken());
-            Log.d("LOGPERFSYNC", userSyncCategoriesModel.getStatus());
         } catch (IOException e) {
             e.printStackTrace();
         }
 
     }
 
-    private void syncExpenses(){
+    private void syncExpenses() {
         RestService restService = new RestService();
         List<ExpensesEntity> expensesEntityList = ExpensesEntity.selectAll("");
         List<ExpensesModel> expensesModels = new ArrayList<>();
-        for (int i = 0; i < expensesEntityList.size(); i ++){
+        for (int i = 0; i < expensesEntityList.size(); i++) {
             ExpensesModel expensesModel = new ExpensesModel();
             expensesModel.setId(Integer.parseInt(expensesEntityList.get(i).getId().toString()));
             expensesModel.setSum(Double.parseDouble(expensesEntityList.get(i).getSum()));
@@ -127,6 +151,7 @@ public class TrackerSyncAdapter extends AbstractThreadedSyncAdapter {
         String gson = new Gson().toJson(expensesModels);
         try {
             UserSyncExpensesModel userSyncExpensesModel = restService.userSyncExpensesModel(gson, MoneyTrackerAplication.getAuthToken(), MoneyTrackerAplication.getGoogleAuthToken());
+
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -141,6 +166,36 @@ public class TrackerSyncAdapter extends AbstractThreadedSyncAdapter {
         } else {
             ContentResolver.addPeriodicSync(account, authority, new Bundle(), syncInterval);
         }
+    }
+
+    public void sendNotification() {
+        init();
+        if (!isNotificationsEnabled) {
+            return;
+        }
+        mNotificationManager = (NotificationManager) getContext().getSystemService(Context.NOTIFICATION_SERVICE);
+
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(getContext())
+
+                .setSmallIcon(R.mipmap.logo_log)
+                .setLargeIcon(BitmapFactory.decodeResource(getContext().getResources(), R.mipmap.logo_log))
+                .setContentTitle(getContext().getString(R.string.app_name))
+                .setContentText(getContext().getString(R.string.notification_message))
+                .setAutoCancel(true);
+
+        if (isLedEnabled) {
+            builder.setLights(Color.BLUE, LED_LIGHTS_TIME_ON, LED_LIGHTS_TIME_OFF);
+        }
+
+        if (isSoundEnabled) {
+            builder.setSound(Settings.System.DEFAULT_NOTIFICATION_URI);
+        }
+
+        if (isVibrateEnabled) {
+            builder.setVibrate(VIBRATE_PATTERN);
+        }
+
+        mNotificationManager.notify(NOTIFICATION_ID, builder.build());
     }
 
     public static void initializeSyncAdapter(Context context) {
