+++ app/src/main/res/layout/activity_main.xml
@@ -36,6 +36,7 @@
         <FrameLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
+            android:layout_marginTop="?attr/actionBarSize"
             android:id="@+id/main_container"
             android:layout_below="@id/app_bar_layout"/>
