+++ app/src/main/res/layout/category_fragment.xml
@@ -9,7 +9,7 @@
     <android.support.v4.widget.SwipeRefreshLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginTop="?attr/actionBarSize"
+
         android:id="@+id/swipe_layout_category">
