+++ app/src/main/res/layout/expenses_fragment.xml
@@ -8,7 +8,7 @@
    <android.support.v4.widget.SwipeRefreshLayout
        android:layout_width="match_parent"
        android:layout_height="wrap_content"
-       android:layout_marginTop="?attr/actionBarSize"
+
        android:id="@+id/swipe_layout">
 
    <android.support.v7.widget.RecyclerView
