+++ app/src/main/res/layout/statistic_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:layout_marginTop="?attr/actionBarSize"
+
     android:orientation="vertical">
     <LinearLayout
         android:layout_width="match_parent"
